Refactor css files for online editors
