+++ client/app/bundles/course/assessment/question/programming/components/OnlineEditorCppView.jsx
@@ -16,7 +16,7 @@ import { grey100, grey300, white } from 'material-ui/styles/colors';
 import 'brace/mode/c_cpp';
 import 'brace/theme/monokai';
 
-import styles from './OnlineEditorCppView.scss';
+import styles from './OnlineEditorView.scss';
 import translations from './OnlineEditorView.intl';
 
 const MAX_TEST_CASES = 99;
